refactor: unify sync and future APIs
Added macros to share the non-blocking implementations for both `Client`
and `Sender` variants.
